requirements.txt
# Core dependencies - optimized for free tiers
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
plotly>=5.15.0
pyyaml>=6.0.0
psutil>=5.9.0
openpyxl>=3.1.0  # Para importación de archivos Excel

# ML dependencies
scikit-learn>=1.3.0
prophet>=1.0.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Automation dependencies - Agents.md compliance
celery>=5.3.0
redis>=4.6.0
websockets>=12.0
python-multipart>=0.0.6

# HTTP client for external integrations
requests>=2.31.0

# Data generation dependencies
Faker>=20.0.0

# Optional - remove if causing deployment issues
# python-dateutil>=2.8.0
# urllib3>=2.0.0
